+++ src/interfaces/PenalCode.ts
@@ -2,4 +2,5 @@ export interface PenalCode {
   id: string;
   title: string;
   des: string;
+  jail_time: string;
 }