+++ src/lib/database.ts
@@ -50,6 +50,8 @@ async function updateLine(sql: string) {
 }
 
 async function updateDb() {
+  updateLine("ALTER TABLE `officers` ADD `suspended` varchar(255) DEFAULT '0' AFTER `status`;");
+  updateLine("ALTER TABLE `penal_codes` ADD `jail_time` varchar(255) DEFAULT NULL AFTER `des`;");
   updateLine(`
   CREATE TABLE \`mugshots\` (
     \`id\` varchar(255) NOT NULL,